Uploaded autodate_input_clean and autodate_input_with_intralatin_derivatives
Both files have a new column for the source of the loanword.
autodate_input_clean includes loanwords that are not: (1) derived from Hebrew, (2) derived within Irish from stems that are loanwords, or (3) derived within Latin. Possible exception: Alpes, alpinus!
autodate_input_with_intralatin_derivatives contains the same loanwords as in the previous file, except that intra-Latin derivatives are included. This includes caindleóir, if caindleóir is thought of as a derivation within Latin.